Merge Upstream

+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/CassandraMailboxMapper.java
@@ -19,16 +19,14 @@
 
 package org.apache.james.mailbox.cassandra.mail;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.StringTokenizer;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
+import java.time.Duration;
 
 import javax.inject.Inject;
 
 import org.apache.commons.lang3.tuple.Pair;
+import org.apache.james.backends.cassandra.versions.CassandraSchemaVersionManager;
+import org.apache.james.backends.cassandra.versions.SchemaVersion;
+import org.apache.james.core.Username;
 import org.apache.james.mailbox.acl.ACLDiff;
 import org.apache.james.mailbox.cassandra.ids.CassandraId;
 import org.apache.james.mailbox.exception.MailboxException;
@@ -39,55 +37,87 @@ import org.apache.james.mailbox.model.MailboxACL;
 import org.apache.james.mailbox.model.MailboxACL.Right;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.mailbox.model.MailboxPath;
+import org.apache.james.mailbox.model.UidValidity;
+import org.apache.james.mailbox.model.search.MailboxQuery;
 import org.apache.james.mailbox.store.mail.MailboxMapper;
+import org.apache.james.util.FunctionalUtils;
 import org.apache.james.util.ReactorUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.github.steveash.guavate.Guavate;
-import com.google.common.collect.ImmutableList;
+import com.google.common.base.Preconditions;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import reactor.util.retry.Retry;
 
 public class CassandraMailboxMapper implements MailboxMapper {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CassandraMailboxMapper.class);
 
-    private static final String WILDCARD = "%";
-    public static final Logger LOGGER = LoggerFactory.getLogger(CassandraMailboxMapper.class);
+    private static final int MAX_RETRY = 5;
+    private static final Duration MIN_RETRY_BACKOFF = Duration.ofMillis(10);
+    private static final Duration MAX_RETRY_BACKOFF = Duration.ofMillis(1000);
+    private static final SchemaVersion MAILBOX_PATH_V_2_MIGRATION_PERFORMED_VERSION = new SchemaVersion(6);
 
     private final CassandraMailboxDAO mailboxDAO;
     private final CassandraMailboxPathDAOImpl mailboxPathDAO;
     private final CassandraMailboxPathV2DAO mailboxPathV2DAO;
     private final CassandraACLMapper cassandraACLMapper;
     private final CassandraUserMailboxRightsDAO userMailboxRightsDAO;
+    private final CassandraSchemaVersionManager versionManager;
 
     @Inject
-    public CassandraMailboxMapper(CassandraMailboxDAO mailboxDAO, CassandraMailboxPathDAOImpl mailboxPathDAO, CassandraMailboxPathV2DAO mailboxPathV2DAO, CassandraUserMailboxRightsDAO userMailboxRightsDAO, CassandraACLMapper aclMapper) {
+    public CassandraMailboxMapper(CassandraMailboxDAO mailboxDAO,
+                                  CassandraMailboxPathDAOImpl mailboxPathDAO,
+                                  CassandraMailboxPathV2DAO mailboxPathV2DAO,
+                                  CassandraUserMailboxRightsDAO userMailboxRightsDAO,
+                                  CassandraACLMapper aclMapper,
+                                  CassandraSchemaVersionManager versionManager) {
         this.mailboxDAO = mailboxDAO;
         this.mailboxPathDAO = mailboxPathDAO;
         this.mailboxPathV2DAO = mailboxPathV2DAO;
         this.userMailboxRightsDAO = userMailboxRightsDAO;
         this.cassandraACLMapper = aclMapper;
+        this.versionManager = versionManager;
+    }
+
+    private Mono<Boolean> needMailboxPathV1Support() {
+        return versionManager.isBefore(MAILBOX_PATH_V_2_MIGRATION_PERFORMED_VERSION);
     }
 
     @Override
-    public void delete(Mailbox mailbox) {
+    public Mono<Void> delete(Mailbox mailbox) {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-        Flux.merge(
-                mailboxPathDAO.delete(mailbox.generateAssociatedPath()),
-                mailboxPathV2DAO.delete(mailbox.generateAssociatedPath()))
-            .thenEmpty(mailboxDAO.delete(mailboxId))
-            .block();
+        return deletePath(mailbox)
+            .thenEmpty(mailboxDAO.delete(mailboxId)
+                .retryWhen(Retry.backoff(MAX_RETRY, MIN_RETRY_BACKOFF).maxBackoff(MAX_RETRY_BACKOFF)));
+    }
+
+    private Flux<Void> deletePath(Mailbox mailbox) {
+        return needMailboxPathV1Support()
+            .flatMapMany(needSupport -> {
+                if (needSupport) {
+                    return Flux.merge(
+                        mailboxPathDAO.delete(mailbox.generateAssociatedPath()),
+                        mailboxPathV2DAO.delete(mailbox.generateAssociatedPath()));
+                }
+                return Flux.from(mailboxPathV2DAO.delete(mailbox.generateAssociatedPath()));
+            });
     }
 
     @Override
-    public Mailbox findMailboxByPath(MailboxPath path) throws MailboxException {
+    public Mono<Mailbox> findMailboxByPath(MailboxPath path) {
         return mailboxPathV2DAO.retrieveId(path)
             .map(CassandraIdAndPath::getCassandraId)
             .flatMap(this::retrieveMailbox)
-            .switchIfEmpty(fromPreviousTable(path))
-            .blockOptional()
-            .orElseThrow(() -> new MailboxNotFoundException(path));
+            .switchIfEmpty(fromPreviousTable(path));
+    }
+
+    @Override
+    public Mono<Boolean> pathExists(MailboxPath mailboxName) {
+        return mailboxPathV2DAO.retrieveId(mailboxName)
+            .switchIfEmpty(mailboxPathDAO.retrieveId(mailboxName))
+            .hasElement();
     }
 
     private Mono<Mailbox> fromPreviousTable(MailboxPath path) {
@@ -114,18 +144,22 @@ public class CassandraMailboxMapper implements MailboxMapper {
     }
 
     @Override
-    public Mailbox findMailboxById(MailboxId id) throws MailboxException {
+    public Mono<Mailbox> findMailboxById(MailboxId id) {
         CassandraId mailboxId = (CassandraId) id;
         return retrieveMailbox(mailboxId)
-            .blockOptional()
-            .orElseThrow(() -> new MailboxNotFoundException(id));
+            .switchIfEmpty(Mono.error(() -> new MailboxNotFoundException(id)));
     }
 
     private Mono<Mailbox> retrieveMailbox(CassandraId mailboxId) {
-        Mono<MailboxACL> acl = cassandraACLMapper.getACL(mailboxId);
+        Mono<MailboxACL> acl = retrieveAcl(mailboxId);
         Mono<Mailbox> simpleMailbox = mailboxDAO.retrieveMailbox(mailboxId);
 
         return acl.zipWith(simpleMailbox, this::addAcl);
+    }
+
+    private Mono<MailboxACL> retrieveAcl(CassandraId mailboxId) {
+        return cassandraACLMapper.getACL(mailboxId)
+            .defaultIfEmpty(MailboxACL.EMPTY);
     }
 
     private Mailbox addAcl(MailboxACL acl, Mailbox mailbox) {
@@ -134,30 +168,25 @@ public class CassandraMailboxMapper implements MailboxMapper {
     }
 
     @Override
-    public List<Mailbox> findMailboxWithPathLike(MailboxPath path) {
-        List<Mailbox> mailboxesV2 = toMailboxes(path, mailboxPathV2DAO.listUserMailboxes(path.getNamespace(), path.getUser()));
-        List<Mailbox> mailboxesV1 = toMailboxes(path, mailboxPathDAO.listUserMailboxes(path.getNamespace(), path.getUser()));
+    public Flux<Mailbox> findMailboxWithPathLike(MailboxQuery.UserBound query) {
+        String fixedNamespace = query.getFixedNamespace();
+        Username fixedUser = query.getFixedUser();
 
-        List<Mailbox> mailboxesV1NotInV2 = mailboxesV1.stream()
-            .filter(mailboxV1 -> mailboxesV2.stream()
-                .map(Mailbox::generateAssociatedPath)
-                .noneMatch(mailboxV2path -> mailboxV2path.equals(mailboxV1.generateAssociatedPath())))
-            .collect(Guavate.toImmutableList());
-
-        return ImmutableList.<Mailbox>builder()
-            .addAll(mailboxesV2)
-            .addAll(mailboxesV1NotInV2)
-            .build();
+        return listPaths(fixedNamespace, fixedUser)
+            .filter(idAndPath -> query.isPathMatch(idAndPath.getMailboxPath()))
+            .distinct(CassandraIdAndPath::getMailboxPath)
+            .concatMap(this::retrieveMailbox);
     }
 
-    private List<Mailbox> toMailboxes(MailboxPath path, Flux<CassandraIdAndPath> listUserMailboxes) {
-        Pattern regex = Pattern.compile(constructEscapedRegexForMailboxNameMatching(path));
-
-        return listUserMailboxes
-                .filter(idAndPath -> regex.matcher(idAndPath.getMailboxPath().getName()).matches())
-                .flatMap(this::retrieveMailbox)
-                .collectList()
-                .block();
+    private Flux<CassandraIdAndPath> listPaths(String fixedNamespace, Username fixedUser) {
+        return needMailboxPathV1Support()
+            .flatMapMany(needSupport -> {
+                if (needSupport) {
+                    return Flux.concat(mailboxPathV2DAO.listUserMailboxes(fixedNamespace, fixedUser),
+                        mailboxPathDAO.listUserMailboxes(fixedNamespace, fixedUser));
+                }
+                return mailboxPathV2DAO.listUserMailboxes(fixedNamespace, fixedUser);
+            });
     }
 
     private Mono<Mailbox> retrieveMailbox(CassandraIdAndPath idAndPath) {
@@ -167,50 +196,66 @@ public class CassandraMailboxMapper implements MailboxMapper {
     }
 
     @Override
-    public MailboxId save(Mailbox mailbox) throws MailboxException {
-        CassandraId cassandraId = retrieveId(mailbox);
-        mailbox.setMailboxId(cassandraId);
-        if (!trySave(mailbox, cassandraId)) {
-            throw new MailboxExistsException(mailbox.generateAssociatedPath().asString());
-        }
-        return cassandraId;
-    }
+    public Mono<Mailbox> create(MailboxPath mailboxPath, UidValidity uidValidity) {
+        CassandraId cassandraId = CassandraId.timeBased();
+        Mailbox mailbox = new Mailbox(mailboxPath, uidValidity, cassandraId);
 
-    private boolean trySave(Mailbox cassandraMailbox, CassandraId cassandraId) {
-        boolean isCreated = mailboxPathV2DAO.save(cassandraMailbox.generateAssociatedPath(), cassandraId).block();
-        if (isCreated) {
-            Optional<Mailbox> simpleMailbox = retrieveMailbox(cassandraId).blockOptional();
-            simpleMailbox.ifPresent(mbx -> mailboxPathV2DAO.delete(mbx.generateAssociatedPath()).block());
-            mailboxDAO.save(cassandraMailbox).block();
-        }
-        return isCreated;
-    }
-
-    private CassandraId retrieveId(Mailbox cassandraMailbox) {
-        if (cassandraMailbox.getMailboxId() == null) {
-            return CassandraId.timeBased();
-        } else {
-            return (CassandraId) cassandraMailbox.getMailboxId();
-        }
+        return mailboxPathV2DAO.save(mailbox.generateAssociatedPath(), cassandraId)
+            .filter(isCreated -> isCreated)
+            .flatMap(mailboxHasCreated -> persistMailboxEntity(mailbox)
+                .thenReturn(mailbox))
+            .switchIfEmpty(Mono.error(() -> new MailboxExistsException(mailbox.generateAssociatedPath().asString())));
     }
 
     @Override
-    public boolean hasChildren(Mailbox mailbox, char delimiter) {
+    public Mono<MailboxId> rename(Mailbox mailbox) {
+        Preconditions.checkNotNull(mailbox.getMailboxId(), "A mailbox we want to rename should have a defined mailboxId");
+
+        CassandraId cassandraId = (CassandraId) mailbox.getMailboxId();
+        return tryRename(mailbox, cassandraId)
+            .filter(FunctionalUtils.identityPredicate())
+            .switchIfEmpty(Mono.error(() -> new MailboxExistsException(mailbox.generateAssociatedPath().asString())))
+            .thenReturn(cassandraId);
+    }
+
+    private Mono<Boolean> tryRename(Mailbox cassandraMailbox, CassandraId cassandraId) {
+        return mailboxDAO.retrieveMailbox(cassandraId)
+            .flatMap(mailbox -> mailboxPathV2DAO.save(cassandraMailbox.generateAssociatedPath(), cassandraId)
+                .filter(isCreated -> isCreated)
+                .flatMap(mailboxHasCreated -> deletePreviousMailboxPathReference(mailbox.generateAssociatedPath())
+                    .then(persistMailboxEntity(cassandraMailbox))
+                    .thenReturn(true))
+                .defaultIfEmpty(false))
+            .switchIfEmpty(Mono.error(() -> new MailboxNotFoundException(cassandraId)));
+    }
+
+    private Mono<Void> persistMailboxEntity(Mailbox cassandraMailbox) {
+        return mailboxDAO.save(cassandraMailbox)
+            .retryWhen(Retry.backoff(MAX_RETRY, MIN_RETRY_BACKOFF).maxBackoff(MAX_RETRY_BACKOFF));
+    }
+
+    private Mono<Void> deletePreviousMailboxPathReference(MailboxPath mailboxPath) {
+        return mailboxPathV2DAO.delete(mailboxPath)
+            .retryWhen(Retry.backoff(MAX_RETRY, MIN_RETRY_BACKOFF).maxBackoff(MAX_RETRY_BACKOFF));
+    }
+
+    @Override
+    public Mono<Boolean> hasChildren(Mailbox mailbox, char delimiter) {
         return Flux.merge(
                 mailboxPathDAO.listUserMailboxes(mailbox.getNamespace(), mailbox.getUser()),
                 mailboxPathV2DAO.listUserMailboxes(mailbox.getNamespace(), mailbox.getUser()))
-            .filter(idAndPath -> idAndPath.getMailboxPath().getName().startsWith(mailbox.getName() + String.valueOf(delimiter)))
-            .hasElements()
-            .block();
+            .filter(idAndPath -> isPathChildOfMailbox(idAndPath, mailbox, delimiter))
+            .hasElements();
+    }
+
+    private boolean isPathChildOfMailbox(CassandraIdAndPath idAndPath, Mailbox mailbox, char delimiter) {
+        return idAndPath.getMailboxPath().getName().startsWith(mailbox.getName() + String.valueOf(delimiter));
     }
 
     @Override
-    public List<Mailbox> list() {
+    public Flux<Mailbox> list() {
         return mailboxDAO.retrieveAllMailboxes()
-            .flatMap(this::toMailboxWithAcl)
-            .map(simpleMailboxes -> (Mailbox) simpleMailboxes)
-            .collectList()
-            .block();
+            .flatMap(this::toMailboxWithAcl);
     }
 
     @Override
@@ -219,13 +264,13 @@ public class CassandraMailboxMapper implements MailboxMapper {
     }
 
     @Override
-    public ACLDiff updateACL(Mailbox mailbox, MailboxACL.ACLCommand mailboxACLCommand) throws MailboxException {
+    public Mono<ACLDiff> updateACL(Mailbox mailbox, MailboxACL.ACLCommand mailboxACLCommand) {
         CassandraId cassandraId = (CassandraId) mailbox.getMailboxId();
         return cassandraACLMapper.updateACL(cassandraId, mailboxACLCommand);
     }
 
     @Override
-    public ACLDiff setACL(Mailbox mailbox, MailboxACL mailboxACL) throws MailboxException {
+    public Mono<ACLDiff> setACL(Mailbox mailbox, MailboxACL mailboxACL) {
         CassandraId cassandraId = (CassandraId) mailbox.getMailboxId();
         return cassandraACLMapper.setACL(cassandraId, mailboxACL);
     }
@@ -235,25 +280,9 @@ public class CassandraMailboxMapper implements MailboxMapper {
         // Do nothing
     }
 
-    private String constructEscapedRegexForMailboxNameMatching(MailboxPath path) {
-        return Collections
-            .list(new StringTokenizer(path.getName(), WILDCARD, true))
-            .stream()
-            .map(this::tokenToPatternPart)
-            .collect(Collectors.joining());
-    }
-
-    private String tokenToPatternPart(Object token) {
-        if (token.equals(WILDCARD)) {
-            return ".*";
-        } else {
-            return Pattern.quote((String) token);
-        }
-    }
-
     private Mono<Mailbox> toMailboxWithAcl(Mailbox mailbox) {
         CassandraId cassandraId = (CassandraId) mailbox.getMailboxId();
-        return cassandraACLMapper.getACL(cassandraId)
+        return retrieveAcl(cassandraId)
             .map(acl -> {
                 mailbox.setACL(acl);
                 return mailbox;
@@ -261,18 +290,10 @@ public class CassandraMailboxMapper implements MailboxMapper {
     }
 
     @Override
-    public List<Mailbox> findNonPersonalMailboxes(String userName, Right right) {
+    public Flux<Mailbox> findNonPersonalMailboxes(Username userName, Right right) {
         return userMailboxRightsDAO.listRightsForUser(userName)
-            .filter(mailboxId -> authorizedMailbox(mailboxId.getRight(), right))
+            .filter(mailboxId -> mailboxId.getRight().contains(right))
             .map(Pair::getLeft)
-            .flatMap(this::retrieveMailbox)
-            .map(simpleMailboxes -> (Mailbox) simpleMailboxes)
-            .collectList()
-            .block();
+            .flatMap(this::retrieveMailbox);
     }
-
-    private boolean authorizedMailbox(MailboxACL.Rfc4314Rights rights, Right right) {
-        return rights.contains(right);
-    }
-
 }
